+++ src/DetailsView/reports/components/report-sections/collapsible-container.tsx
